+++ CHANGES.rst
@@ -1,4 +1,4 @@
-Changes in Riot 0.8.22 (2019-XX-XX)
+Changes in Riot 0.8.23 (2019-XX-XX)
 ===================================================
 
 MatrixSdk:
@@ -6,6 +6,34 @@ MatrixSdk:
 
 Features:
  -
+
+Improvements:
+ -
+
+Other changes:
+ -
+
+Bugfix:
+ -
+
+Translations:
+ -
+
+Build:
+ -
+
+Changes in Riot 0.8.22 (2019-02-01)
+===================================================
+
+MatrixSdk:
+ - Upgrade MatrixSdk to version 0.9.16.
+
+Features:
+ - keys backup: Implement setup screen (#2883)
+ - keys backup: Display a warning on new sign out screen (#2885)
+ - keys backup: recover screen (#2887)
+ - keys backup: Add a dedicated section to settings (#2884)
+ - keys backup: Add a banner on Home to setup or recover backup (#2884)
 
 Improvements:
  - Support Split-screen mode (#1832)
@@ -18,11 +46,12 @@ Improvements:
  - Kotlin Code Improvement in VectorSettingsPreferencesFragment.kt
  - Remove redundant !! , Replace it with null safe operators in VectorSettingsPreferencesFragment.kt
  - `Redact` has been renamed to `Remove` to match riot/web (#2871)
-
+ - Remove long click download action in MediaViewer (#2882)
 
 Other changes:
  - Update of Light and Dark themes (#2710)
  - Restore the crash report dialog after a crash
+ - New application icon! (#2905)
 
 Bugfix:
  - Fix crash when taking picture for user avatar on old device (#2818)
@@ -31,15 +60,11 @@ Bugfix:
  - FCM Troubleshoot screen crash in some cases (#2846)
  - Fix login button issue (#1568)
  - Fix issue with registration when an email is provided (#2852)
+ - Fix issues with Tombstone events (#2866 && #2867)
  - Fix crash on BugReportActivity if previous Activity is destroyed (#2876)
+ - Key share request does not go away when user select "verify" (#2781)
 
-Translations:
- -
-
-Build:
- -
-
-Changes in Riot 0.8.21 (2018-01-02)
+Changes in Riot 0.8.21 (2019-01-02)
 ===================================================
 
 MatrixSdk: